# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HR Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /attendance/me:
    get:
      operationId: getMyAttendance
      tags: [attendance]
      summary: Get the current user's attendance history
      parameters:
        - name: month
          in: query
          required: false
          schema: { type: string, description: "YYYY-MM" }
      responses:
        "200":
          description: Attendance history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
  /attendance/employee/{id}:
    get:
      operationId: getEmployeeAttendance
      tags: [attendance]
      summary: Get an employee's attendance history (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: month
          in: query
          required: false
          schema: { type: string, description: "YYYY-MM" }
      responses:
        "200":
          description: Attendance history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
  /attendance/today-summary:
    get:
      operationId: getTodayAttendanceSummary
      tags: [attendance, dashboard]
      summary: Get attendance summary for all employees on a date (admin). Defaults to today.
      parameters:
        - name: date
          in: query
          required: false
          schema: { type: string, description: "YYYY-MM-DD; defaults to today" }
      responses:
        "200":
          description: Attendance summary for the date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceSummary"
  /attendance/calendar:
    get:
      operationId: getAttendanceCalendar
      tags: [attendance]
      summary: Get an employee's attendance for a month formatted for the calendar (admin only)
      parameters:
        - name: employeeId
          in: query
          required: true
          schema: { type: integer }
        - name: month
          in: query
          required: false
          schema: { type: string, description: "YYYY-MM" }
      responses:
        "200":
          description: Calendar data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceCalendar"
  /attendance/override:
    post:
      operationId: overrideAttendance
      tags: [attendance]
      summary: Manually set attendance status for an employee on a date (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceOverrideRequest"
      responses:
        "200":
          description: Updated record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /attendance/{id}/excuse:
    post:
      operationId: setAttendanceExcused
      tags: [attendance]
      summary: Mark a late/half-day attendance row as excused or not excused (admin/HR only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetAttendanceExcusedRequest"
      responses:
        "200":
          description: Updated record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /attendance/employee/{id}/lates:
    get:
      operationId: getEmployeeLateRecords
      tags: [attendance]
      summary: List an employee's late attendance rows in a month (admin/HR only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: month
          in: query
          required: false
          schema: { type: string, description: "YYYY-MM; defaults to current month" }
      responses:
        "200":
          description: Late attendance rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
...
components:
  schemas:
...
    AttendanceRecord:
      type: object
      required: [id, employeeId, date, status]
      properties:
        id: { type: integer }
        employeeId: { type: integer }
        employeeName: { type: string }
        date: { type: string, format: date }
        checkInTime: { type: string, nullable: true, description: "ISO datetime" }
        checkOutTime: { type: string, nullable: true, description: "ISO datetime" }
        workedMinutes: { type: integer, nullable: true }
        status:
          type: string
          enum: [present, late, absent, on_leave, half_day, remote_work]
        isLate: { type: boolean }
        excused: { type: boolean, description: "True when an approved request (late/half-day) means this row should not deduct from payroll" }
        notes: { type: string, nullable: true }
...
    AttendanceSummary:
      type: object
      required:
        - date
        - total
        - present
        - late
        - absent
        - onLeave
        - halfDay
        - remoteWork
      properties:
        date: { type: string, format: date }
        total: { type: integer }
        present: { type: integer }
        late: { type: integer }
        absent: { type: integer }
        onLeave: { type: integer }
        halfDay: { type: integer }
        remoteWork: { type: integer }
        records:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceRecord"
    AttendanceCalendarDay:
      type: object
      required: [date, status]
      properties:
        date: { type: string, format: date }
        status:
          type: string
          enum: [present, late, absent, on_leave, half_day, remote_work, weekend, future, none]
        record:
          oneOf:
            - $ref: "#/components/schemas/AttendanceRecord"
            - type: "null"
    AttendanceCalendar:
      type: object
      required: [employeeId, employeeName, month, days]
      properties:
        employeeId: { type: integer }
        employeeName: { type: string }
        month: { type: string, description: "YYYY-MM" }
        days:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceCalendarDay"
    AttendanceOverrideRequest:
      type: object
      required: [employeeId, date, status]
      properties:
        employeeId: { type: integer }
        date: { type: string, format: date }
        status:
          type: string
          enum: [present, late, absent, on_leave, half_day, remote_work]
        notes: { type: string, nullable: true }
    SetAttendanceExcusedRequest:
      type: object
      required: [excused]
      properties:
        excused: { type: boolean }
...